Refactor CharacterGroupOverlay to BulkEditOverlay
